requirements.txt
py-cord
discord.py
dotenv